Merge RE patches

+++ brut.j.util/src/main/java/brut/util/OS.java
@@ -17,6 +17,8 @@
 package brut.util;
 
 import brut.common.BrutException;
+import org.apache.commons.io.IOUtils;
+
 import java.io.*;
 import java.util.Arrays;
 import java.util.concurrent.ExecutorService;
@@ -24,11 +26,6 @@ import java.util.concurrent.Executors;
 import java.util.concurrent.TimeUnit;
 import java.util.logging.Logger;
 
-import org.apache.commons.io.IOUtils;
-
-/**
- * @author Ryszard Wiśniewski <brut.alll@gmail.com>
- */
 public class OS {
 
     private static final Logger LOGGER = Logger.getLogger("");
@@ -38,8 +35,11 @@ public class OS {
             return;
         }
         File[] files = dir.listFiles();
-        for (int i = 0; i < files.length; i++) {
-            File file = files[i];
+        if (files == null) {
+            return;
+        }
+
+        for (File file : files) {
             if (file.isDirectory()) {
                 rmdir(file);
             } else {
@@ -49,7 +49,7 @@ public class OS {
         dir.delete();
     }
         
-    public static void rmfile(String file) throws BrutException {
+    public static void rmfile(String file) {
     	File del = new File(file);
     	del.delete();
     }
@@ -61,10 +61,12 @@ public class OS {
     public static void cpdir(File src, File dest) throws BrutException {
         dest.mkdirs();
         File[] files = src.listFiles();
-        for (int i = 0; i < files.length; i++) {
-            File file = files[i];
-            File destFile = new File(dest.getPath() + File.separatorChar
-                + file.getName());
+        if (files == null) {
+            return;
+        }
+
+        for (File file : files) {
+            File destFile = new File(dest.getPath() + File.separatorChar + file.getName());
             if (file.isDirectory()) {
                 cpdir(file, destFile);
                 continue;
@@ -81,21 +83,21 @@ public class OS {
         }
     }
 
-    public static void cpdir(String src, String dest) throws BrutException {
-        cpdir(new File(src), new File(dest));
-    }
-
     public static void exec(String[] cmd) throws BrutException {
-        Process ps = null;
-        int exitValue = -99;
+        Process ps;
+        int exitValue;
+
         try {
             ProcessBuilder builder = new ProcessBuilder(cmd);
             ps = builder.start();
+
             new StreamForwarder(ps.getErrorStream(), "ERROR").start();
             new StreamForwarder(ps.getInputStream(), "OUTPUT").start();
+
             exitValue = ps.waitFor();
-            if (exitValue != 0)
+            if (exitValue != 0) {
                 throw new BrutException("could not exec (exit code = " + exitValue + "): " + Arrays.toString(cmd));
+            }
         } catch (IOException ex) {
             throw new BrutException("could not exec: " + Arrays.toString(cmd), ex);
         } catch (InterruptedException ex) {
